RPMS/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
python-engineio==3.9.3
python-socketio==4.3.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.15.6
